requirements.txt
pyTelegramBotAPI
pydub
PyDrive
python-dotenv